[Ramses] Keep set_alpha_values in f32 and extend the Epstein CI example
Widening set_alpha_values to f64 shifted the dustywave_godunov.py reference
values by ~5e-8: its rates (1/0.215443, 1/0.464159) are not f32 representable,
unlike dustybox's 100 and 500. That precision fix is orthogonal to Epstein
drag and needs the references regenerated, so revert it and document why the
truncation is there.

The dusty box example now also runs at twice the gas density, where the
Epstein rates must double on their own while the constant rates do not. Both
densities reproduce the constant rate run exactly, which is what checks that
the drag rate is derived from the local gas state rather than from config.

+++ src/shammodels/ramses/include/shammodels/ramses/SolverConfig.hpp
@@ -234,8 +234,12 @@ struct shammodels::basegodunov::SolverConfig {
      *
      * alphas is the dust collision rate (the inverse of the stopping time). Calling this
      * switches the drag config back to DragConfig::ConstantAlphas if it was set otherwise.
+     *
+     * @note the f32 argument truncates the drag rate to single precision. This is kept as is
+     * because the reference values of examples/tests_ci/dustywave_godunov.py were produced with
+     * it, and widening it to f64 shifts them by ~5e-8, well above their 1e-11 margins.
      */
-    inline void set_alphas_static(f64 alpha_values) {
+    inline void set_alphas_static(f32 alpha_values) {
         StackEntry stack_lock{};
         if (!std::holds_alternative<DragConfig::ConstantAlphas>(drag_config.alpha_mode)) {
             drag_config.alpha_mode = DragConfig::ConstantAlphas{};
